Fondle linters balls a bit
