+++ thinking_reflection/model/classes.py
@@ -7,6 +7,7 @@ from thinking_reflection.model.members import FieldDescriptor, MethodDescriptor
 
 def unique_concat[T](*ts: tuple[T, ...]) -> tuple[T, ...]:
     yielded = set()
+
     def i():
         for t in ts:
             for x in t:
@@ -15,6 +16,7 @@ def unique_concat[T](*ts: tuple[T, ...]) -> tuple[T, ...]:
                     yield x
     return tuple(i())
 
+
 def unique_merge[T](d1: frozendict[str, tuple[T]], d2: frozendict[str, tuple[T]]) -> frozendict[str, tuple[T]]:
     out = dict(d1)
     for k, v in d2.items():
@@ -22,6 +24,7 @@ def unique_merge[T](d1: frozendict[str, tuple[T]], d2: frozendict[str, tuple[T]]
             out[k] = tuple()
         out[k] = unique_concat(out[k], v)
     return frozendict(out)
+
 
 class TypeDescriptor(NamedTuple):
     fields: frozendict[str, tuple[FieldDescriptor]]
